+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Yalc changelog
 
 ## 1.0.0.pre.3
+- use .gitignore if no `files` entry in manifest
 
 ## 1.0.0.pre.2 (2017-04-26)